projects/de5ced30-7a74-4aff-874a-93cddf756071/requirements.txt
fastapi==0.92.0
pydantic==1.10.2
sqlalchemy==1.4.39
pyjwt==2.4.0
